+++ pkg/resource/events-logger-config/events-logger-config.go
@@ -3,7 +3,6 @@ package eventsloggerconfig
 import (
 	"fmt"
 
-	"github.com/pkg/errors"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	capi "sigs.k8s.io/cluster-api/api/v1beta1"
@@ -13,30 +12,19 @@ import (
 
 const (
 	eventsLogggerConfigName = "events-logger-config"
-	grafanaAgentConfigName  = "grafana-agent-config"
 )
 
-func generateEventsLoggerConfig(cluster *capi.Cluster, loggingAgent *common.LoggingAgent, tenants []string, includeNamespaces []string, excludeNamespaces []string, installationName string, insecureCA bool, tracingEnabled bool, tempoURL string) (v1.ConfigMap, error) {
+func generateEventsLoggerConfig(cluster *capi.Cluster, tenants []string, includeNamespaces []string, excludeNamespaces []string, installationName string, insecureCA bool, tracingEnabled bool, tempoURL string) (v1.ConfigMap, error) {
 	var values string
 	var err error
 
-	switch loggingAgent.GetKubeEventsLogger() {
-	case common.EventsLoggerGrafanaAgent:
-		values, err = generateGrafanaAgentConfig(cluster, loggingAgent, includeNamespaces, excludeNamespaces, installationName, insecureCA)
-		if err != nil {
-			return v1.ConfigMap{}, err
-		}
-	case common.EventsLoggerAlloy:
-		values, err = generateAlloyEventsConfig(cluster, includeNamespaces, excludeNamespaces, installationName, insecureCA, tracingEnabled, tempoURL, tenants)
-		if err != nil {
-			return v1.ConfigMap{}, err
-		}
-	default:
-		return v1.ConfigMap{}, errors.Errorf("unsupported events logger %q", loggingAgent.GetKubeEventsLogger())
+	values, err = generateAlloyEventsConfig(cluster, includeNamespaces, excludeNamespaces, installationName, insecureCA, tracingEnabled, tempoURL, tenants)
+	if err != nil {
+		return v1.ConfigMap{}, err
 	}
 
 	configmap := v1.ConfigMap{
-		ObjectMeta: configMeta(cluster, loggingAgent),
+		ObjectMeta: configMeta(cluster),
 		Data: map[string]string{
 			"values": values,
 		},
@@ -46,9 +34,9 @@ func generateEventsLoggerConfig(cluster *capi.Cluster, loggingAgent *common.Logg
 }
 
 // ConfigMeta returns metadata for the logging-config
-func configMeta(cluster *capi.Cluster, loggingAgent *common.LoggingAgent) metav1.ObjectMeta {
+func configMeta(cluster *capi.Cluster) metav1.ObjectMeta {
 	metadata := metav1.ObjectMeta{
-		Name:      getEventsLoggerConfigName(cluster, loggingAgent),
+		Name:      getEventsLoggerConfigName(cluster),
 		Namespace: cluster.GetNamespace(),
 		Labels:    map[string]string{},
 	}
@@ -57,11 +45,6 @@ func configMeta(cluster *capi.Cluster, loggingAgent *common.LoggingAgent) metav1
 	return metadata
 }
 
-func getEventsLoggerConfigName(cluster *capi.Cluster, loggingAgent *common.LoggingAgent) string {
-	switch loggingAgent.GetKubeEventsLogger() {
-	case common.EventsLoggerGrafanaAgent:
-		return fmt.Sprintf("%s-%s", cluster.GetName(), grafanaAgentConfigName)
-	default:
-		return fmt.Sprintf("%s-%s", cluster.GetName(), eventsLogggerConfigName)
-	}
+func getEventsLoggerConfigName(cluster *capi.Cluster) string {
+	return fmt.Sprintf("%s-%s", cluster.GetName(), eventsLogggerConfigName)
 }